+++ CHANGELOG.md
@@ -9,6 +9,16 @@ All notable changes to this project will be documented in this file. The format 
 - Redesign: new design in all the right places. Giving the users a better way of showing the Objectives and Key Results
 - Content Loaders! We have added content loaders to some of our components and hopefully this will increase the UX for our users
 
+## [2.4.0] 2021-10-04
+
+### Added
+
+- Log in with Microsoft (Thank you, @knowit, for the changes)
+
+### Removed
+
+- Removed keycloak integration
+
 ## [2.3.1] 2021-09-28
 
 ### Fixed